+++ backend/.env
@@ -2,5 +2,5 @@ DATABASE_URL="postgresql://postgres.pzzbmwmmtstgswoqibtg:@DangTrang08282024@aws-
 DIRECT_URL="postgresql://postgres.pzzbmwmmtstgswoqibtg:@DangTrang08282024@aws-0-us-east-1.pooler.supabase.com:5432/postgres"
 DEBUG="*"
 
-JWT_SECRET=your_super_secret_key
+JWT_SECRET="mimimeomeo"
 JWT_EXPIRES_IN=1h
