+++ backend/.gitignore
@@ -7,3 +7,4 @@ logs
 npm-debug.log*
 # Ignore package-lock.json
 package-lock.json
+.env
