--- backend/src/routes/User.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const UserRouter = express.Router();
-const {login, signup} = require("../controllers/User")
-
-UserRouter.get('/login', login)
-UserRouter.post('/signup',signup)
-
-module.exports = UserRouter
