--- backend/src/utils/db.js
@@ -1,19 +1,0 @@
-// utils/db.js
-
-const mongoose = require('mongoose');
-require('dotenv').config();
-
-const connectDB = async () => {
-    try {
-        const conn = await mongoose.connect(process.env.MONGO_URI, {
-            useNewUrlParser: true,
-            useUnifiedTopology: true,
-        });
-        console.log(`MongoDB connected: ${conn.connection.host}`);
-    } catch (error) {
-        console.error(`Error: ${error.message}`);
-        process.exit(1);
-    }
-};
-
-module.exports = connectDB;
