--- frontend/src/components/Budulanding.jsx
@@ -1,4 +1,0 @@
-import Budu from '../assets/budu.png'
-
-<img src = {Budu} alt = 'Budu'/>
-<
