+++ frontend/src/components/landing/Header.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import avatar from '../assets/avatar.png'
+import avatar from '../../assets/avatar.png'
 
 const Header = () => {
   return (
