+++ frontend/src/components/landing/HeroSection.jsx
